Move find_mismatching_footnotes
