+++ src/database/directory.rs
@@ -24,6 +24,7 @@ impl Directory {
             doc.find_mismatching_sections(&doc_path, &self.config, issues);
             doc.find_unordered_sections(&doc_path, &self.config, issues);
             doc.find_empty_titles(&doc_path, issues);
+            doc.find_mismatching_footnotes(&doc_path, issues);
         }
         for (path, dir) in &self.dirs {
             dir.check(&relative_path.join(path), issues);
