+++ src/scan/mod.rs
@@ -1,7 +1,6 @@
 //! The probes module contains the code that scans for problems in a `TikiBase`
 //! and reports problems found as `Issues`.
 
-pub(crate) mod footnotes;
 pub(crate) mod image_orphaned;
 pub(crate) mod links;
 pub(crate) mod occurrences;
